+++ package.json
@@ -42,6 +42,7 @@
     "rehype-autolink-headings": "^7.1.0",
     "rehype-slug": "^6.0.0",
     "remark": "^15.0.1",
+    "remark-emoji": "^5.0.1",
     "remark-gfm": "^4.0.0",
     "remark-rehype": "^11.1.1",
     "remark-toc": "^9.0.0"
