+++ src/app/docs/components/search/search.module.scss
@@ -1,9 +1,3 @@
-.results {
-  p {
-    color: #555;
-  }
-}
-
 .subresults {
   margin-top: 10px;
   padding-left: 15px;
@@ -16,10 +10,8 @@
 
 .subresult h4 {
   margin: 0;
-  font-size: 1.1em;
 }
 
 .subresult p {
-  font-size: 0.9em;
-  color: #555;
+  font-size: 14px;
 }
