+++ src/core/styles/_constants.scss
@@ -40,6 +40,8 @@ $color-purple-50: hsl(260deg 47% 96%); // #f4f1fa
 // Blue colors
 $color-mentorship-banner: hsl(198deg 83% 70%); // #73ccf2
 $color-mentorship-header: hsl(198deg 86% 55%); // #2BB4EF
+$color-blue-800: hsl(212deg 100% 32%); // #004CA3
+$color-blue-background-100: hsl(212deg 100% calc(45% + 49%) / 1); ; // #e0efff
 
 // Technologies theme colors
 $color-js-bg: hsl(52deg 100% 96%); // #FFFCE9
