+++ src/core/styles/index.scss
@@ -166,7 +166,7 @@ figure {
 
   h2:hover {
     .icon-link::after {
-      content: url('./link.svg');
+      content: url('../../shared/assets/svg/link.svg');
     }
   }
 }
