+++ src/shared/helpers/isValidUrl.ts
@@ -3,6 +3,7 @@ export function isValidUrl(str: string) {
 
   try {
     url = new URL(str);
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   } catch (_) {
     return false;
   }
